# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend de l'application NoLedge, une plateforme de quiz en ligne",
        "title": "NoLedge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/question/all": {
            "get": {
                "description": "Récupère la liste de toutes les questions disponibles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Récupérer toutes les questions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/question.Question"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crée une nouvelle question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Créer une nouvelle question",
                "parameters": [
                    {
                        "description": "Informations de la question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.Question"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/question.Question"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supprime une question et toutes ses propositions associées",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Supprimer une question",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la question",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/get": {
            "get": {
                "description": "Récupère les détails d'une question spécifique",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Récupérer une question par son ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la question",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.Question"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/question/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations d'une question existante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "Mettre à jour une question",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la question",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Nouvelles informations de la question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/question.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.Question"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "proposition.Proposition": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_question": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "question.Question": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_quiz": {
                    "type": "integer"
                },
                "id_type": {
                    "type": "integer"
                },
                "propositions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/proposition.Proposition"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type_question": {
                    "$ref": "#/definitions/question.TypeQuestion"
                }
            }
        },
        "question.TypeQuestion": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}